feat: provide chainId for NFTs
